Added animatedEditTextFocus to Utils.kt
