+++ app/src/main/java/com/example/blanball/presentation/fragments/ResetPasswordStep1Fragment.kt
@@ -1,6 +1,5 @@
 package com.example.blanball.presentation.fragments
 
-import android.animation.ValueAnimator
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
@@ -10,6 +9,7 @@ import androidx.navigation.fragment.findNavController
 import com.example.blanball.R
 import com.example.blanball.databinding.FragmentResetPasswordStep1Binding
 import com.example.blanball.presentation.viewmodels.ResetPasswordStep1ViewModel
+import com.example.blanball.utils.Utils
 
 class ResetPasswordStep1Fragment : Fragment() {
 
@@ -44,24 +44,7 @@ class ResetPasswordStep1Fragment : Fragment() {
             binding.editText5Edit,
         )
 
-        var isFirstFocus = true
-
-        for (i in editTextArray) {
-            i.setOnFocusChangeListener{ _, hasFocus ->
-                if (hasFocus && isFirstFocus){
-                    val anim = ValueAnimator.ofInt(120, 0)
-                    anim.duration = 1000
-                    anim.addUpdateListener { valueAnimator ->
-                        val layoutParams =
-                            binding.mainContainer.layoutParams as ViewGroup.MarginLayoutParams
-                        layoutParams.topMargin = valueAnimator.animatedValue as Int
-                        binding.mainContainer.layoutParams = layoutParams
-                    }
-                    anim.start()
-                    isFirstFocus = false
-                }
-            }
-        }
+        Utils.animateEditTextFocus(editTextArray, binding.mainContainer)
     }
