+++ app/src/main/java/com/example/blanball/presentation/fragments/ResetPasswordStep3Fragment.kt
@@ -1,6 +1,5 @@
 package com.example.blanball.presentation.fragments
 
-import android.animation.ValueAnimator
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
@@ -10,6 +9,7 @@ import androidx.navigation.fragment.findNavController
 import com.example.blanball.R
 import com.example.blanball.databinding.FragmentResetPasswordStep3Binding
 import com.example.blanball.presentation.viewmodels.ResetPasswordStep3ViewModel
+import com.example.blanball.utils.Utils
 
 class ResetPasswordStep3Fragment : Fragment() {
 
@@ -38,24 +38,7 @@ class ResetPasswordStep3Fragment : Fragment() {
            binding.repeatPasswordPlaceholderEdit
        )
 
-        var isFirstFocus = true
-
-        for (i in editTextArray) {
-            i.setOnFocusChangeListener{ _, hasFocus ->
-                if (hasFocus && isFirstFocus){
-                    val anim = ValueAnimator.ofInt(120, 0)
-                    anim.duration = 1000
-                    anim.addUpdateListener { valueAnimator ->
-                        val layoutParams =
-                            binding.mainContainer.layoutParams as ViewGroup.MarginLayoutParams
-                        layoutParams.topMargin = valueAnimator.animatedValue as Int
-                        binding.mainContainer.layoutParams = layoutParams
-                    }
-                    anim.start()
-                    isFirstFocus = false
-                }
-            }
-        }
+        Utils.animateEditTextFocus(editTextArray, binding.mainContainer)
     }
 
     fun navigateToLoginScreen() {
